### 2021-02-07 #3017 @whedon
[PRE REVIEW]: Logodetect: One-shot detection of logos in image and video data
**Submitting author:** @maxpumperla (<a href="http://orcid.org/0000-0002-7801-4184">Max Pumperla</a>)
**Repository:** <a href="https://github.com/Heldenkombinat/Logodetect" target ="_blank">https://github.com/Heldenkombinat/Logodetect</a>
**Version:** v0.1
**Reviewers:** nsaunier, abhishekbajpayee
