+++ paper.bib
@@ -6,6 +6,7 @@
 	publisher = {Curran Associates, Inc.},
 	title = {Faster R-CNN: Towards Real-Time Object Detection with Region Proposal Networks},
 	url = {https://proceedings.neurips.cc/paper/2015/file/14bfa6bb14875e45bba028a21ed38046-Paper.pdf},
+	doi = {10.1109/tpami.2016.2577031},
 	volume = {28},
 	year = {2015}
 }
@@ -45,17 +46,17 @@
 
 
 @article{BHUNIA2019106965,
-	title = "A deep one-shot network for query-based logo retrieval",
-	journal = "Pattern Recognition",
-	volume = "96",
-	pages = "106965",
-	year = "2019",
-	issn = "0031-3203",
-	doi = "https://doi.org/10.1016/j.patcog.2019.106965",
-	url = "http://www.sciencedirect.com/science/article/pii/S0031320319302626",
-	author = "Ayan Kumar Bhunia and Ankan Kumar Bhunia and Shuvozit Ghose and Abhirup Das and Partha Pratim Roy and Umapada Pal",
-	keywords = "Logo retrieval, One-shot learning, Multi-scale conditioning, Similarity matching, Query retrieval",
-	abstract = "Logo detection in real-world scene images is an important problem with applications in advertisement and marketing. Existing general-purpose object detection methods require large training data with annotations for every logo class. These methods do not satisfy the incremental demand of logo classes necessary for practical deployment since it is practically impossible to have such annotated data for new unseen logo. In this work, we develop an easy-to-implement query-based logo detection and localization system by employing a one-shot learning technique using off the shelf neural network components. Given an image of a query logo, our model searches for logo within a given target image and predicts the possible location of the logo by estimating a binary segmentation mask. The proposed model consists of a conditional branch and a segmentation branch. The former gives a conditional latent representation of the given query logo which is combined with feature maps of the segmentation branch at multiple scales in order to obtain the matching location of the query logo in a target image. Feature matching between the latent query representation and multi-scale feature maps of segmentation branch using simple concatenation operation followed by 1 × 1 convolution layer makes our model scale-invariant. Despite its simplicity, our query-based logo retrieval framework achieved superior performance in FlickrLogos-32 and TopLogos-10 dataset over different existing baseline methods."
+	title = {A deep one-shot network for query-based logo retrieval},
+	journal = {Pattern Recognition},
+	volume = {96},
+	pages = {106965},
+	year = {2019},
+	issn = {0031-3203},
+	doi = {10.1016/j.patcog.2019.106965},
+	url = {http://www.sciencedirect.com/science/article/pii/S0031320319302626},
+	author = {Ayan Kumar Bhunia and Ankan Kumar Bhunia and Shuvozit Ghose and Abhirup Das and Partha Pratim Roy and Umapada Pal},
+	keywords = {Logo retrieval, One-shot learning, Multi-scale conditioning, Similarity matching, Query retrieval},
[…]
