### 2021-02-13 #3017 @arokem
Hello @maxpumperla. Thank you for this submission. Could you please say a little bit about the anticipated research uses of this software? I am not sure that this falls within our [definition of research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software)
