### 2021-02-14 #3017 @maxpumperla
Hi @arokem, thanks for taking a first look. We read through your documentation, of course, and were fairly sure that making object detection and recognition more accessible is `solve[ing] complex modeling problems in a scientific context (engineering, ML)`, especially when comparing to other submissions (e.g. LOC and scope).

Specifically, it not easy to get your hands on one-shot detectors and even researchers will have to design them mostly from scratch. @jorgedch and I are currently looking into

- adding more architectures to logodetect for comparison
- extending the problem from logos to other objects, within the same framework
- apply the techniques to concrete problems to see how they fare against established methods (e.g. in banner ads on TV, which doesn't have such sophisticated methods).

The first adds more convenience, the second is "proper" ML research, and the last point is applied science. Now, I can't speak for the community as a whole and what can be anticipated, but the point of OSS for me is that you share first and see if it gets adopted later. Right now I don't see any frameworks allowing you to run and build on top of one-shot detection, and logodetect fills that gap.
