### 2021-03-01 #3017 @arokem
👋 @nsaunier : would you be willing to review this paper for JOSS? Thanks!
