### 2021-03-08 #3017 @nsaunier
> wave @nsaunier : would you be willing to review this paper for JOSS? Thanks!

Yes. Is that the correct way to accept and how should I submit the review? Sorry, I am a beginner here.
