### 2021-03-21 #3124 @whedon
[REVIEW]: Logodetect: One-shot detection of logos in image and video data
**Submitting author:** @maxpumperla (<a href="http://orcid.org/0000-0002-7801-4184">Max Pumperla</a>)
**Repository:** https://github.com/Heldenkombinat/Logodetect
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
