### 2021-03-21 #3124 @arokem
@maxpumperla : a couple of initial comments from me:

- Could you please add more information to the reference "Siamese neural networks for one-shot image recognition."? That reference currently only has authors, year and title. Is there a journal or conference? DOI?
- Is `mtorch` a dependency of the software? Doesn't seem so, but I just wanted to make sure.
