### 2021-03-23 7e683fa
Added more info in paper reference
