+++ paper/paper.bib
@@ -33,17 +33,15 @@
 }
 
 @inproceedings{Koch2015SiameseNN,
-	added-at = {2017-03-27T16:51:39.000+0200},
 	author = {Koch, Gregory and Zemel, Richard and Salakhutdinov, Ruslan},
-	biburl = {https://www.bibsonomy.org/bibtex/26f83b8c4cf316e77e6f6ce1e97411b30/bsc},
-	interhash = {a2596c72ef39240b7c4f86e6038d9344},
-	intrahash = {6f83b8c4cf316e77e6f6ce1e97411b30},
+	url = {https://www.cs.cmu.edu/~rsalakhu/papers/oneshot1.pdf},
 	keywords = {neural-networks one-shot-learning},
-	timestamp = {2017-03-27T16:51:39.000+0200},
 	title = {Siamese Neural Networks for One-shot Image Recognition},
-	year = 2015
+	booktitle = {Proceedings of the 32nd International Conference on Machine Learning (ICML)},
+	publisher = {W&CP},
+    volume = {37},
+    year = {2015},
 }
-
 
 @article{BHUNIA2019106965,
 	title = {A deep one-shot network for query-based logo retrieval},
@@ -56,7 +54,6 @@
 	url = {http://www.sciencedirect.com/science/article/pii/S0031320319302626},
 	author = {Ayan Kumar Bhunia and Ankan Kumar Bhunia and Shuvozit Ghose and Abhirup Das and Partha Pratim Roy and Umapada Pal},
 	keywords = {Logo retrieval, One-shot learning, Multi-scale conditioning, Similarity matching, Query retrieval},
-	abstract = {Logo detection in real-world scene images is an important problem with applications in advertisement and marketing. Existing general-purpose object detection methods require large training data with annotations for every logo class. These methods do not satisfy the incremental demand of logo classes necessary for practical deployment since it is practically impossible to have such annotated data for new unseen logo. In this work, we develop an easy-to-implement query-based logo detection and localization system by employing a one-shot learning technique using off the shelf neural network components. Given an image of a query logo, our model searches for logo within a given target image and predicts the possible location of the logo by estimating a binary segmentation mask. The proposed model consists of a conditional branch and a segmentation branch. The former gives a conditional latent representation of the given query logo which is combined with feature maps of the segmentation branch at multiple scales in order to obtain the matching location of the query logo in a target image. Feature matching between the latent query representation and multi-scale feature maps of segmentation branch using simple concatenation operation followed by 1 × 1 convolution layer makes our model scale-invariant. Despite its simplicity, our query-based logo retrieval framework achieved superior performance in FlickrLogos-32 and TopLogos-10 dataset over different existing baseline methods.}
 }
 
 @article{logo-net,
