### 2021-03-23 #3124 @jorgedch
Hi @arokem, thank you for the feedback.

> * Could you please add more information to the reference "Siamese neural networks for one-shot image recognition."? That reference currently only has authors, year and title. Is there a journal or conference? DOI?

Yes, this is done now. Also I moved the paper files to a folder called _paper/_.

> * Is `mtorch` a dependency of the software? Doesn't seem so, but I just wanted to make sure.

You are right, `mtorch` is not a dependency, we only adapted its structure to our training pipeline.
