### 2021-03-23 #3124 @jorgedch
OK, just noticed that moving the files to _paper/_ breaks the build so I reverted the change.
