### 2021-06-17 #3124 @abhishekbajpayee
@arokem I am unable to edit the checkboxes above even after accepting the invite and being logged in. Can you advise on what I should do?

Until then, some early feedback to @jorgedch:
- Using `pip install` to directly install from pypi takes way too much time. Is there any way to make the install size smaller?
- Provided instructions don't work as-is. The url in the git clone command should ideally be the https one so that anyone is able to clone. The `cd` command right after it has a typo (lowercase l in logodetect).
- On install using `pip install -e` with a local clone, I run into a `No module named 'skbuild'` error.

FYI I am using `python3` and consequently `pip3`.
