### 2021-08-07 #3124 @maxpumperla
@jorgedch any chance we can find the time to address the above findings by @abhishekbajpayee above? Might be good to sync on a round of Q&A together.

@arfon @arokem is this the normal/expected process for a paper submission? I'm just asking because this smaller logodetect paper is a "test run" for more high-level projects that I'm running, and that would fit your journal quite well. I know it's been a strange couple of years with Covid, but it'd be great to have some sort of process and reliability.
