### 2021-08-07 #3124 @arfon
> @arfon @arokem is this the normal/expected process for a paper submission? I'm just asking because this smaller logodetect paper is a "test run" for more high-level projects that I'm running, and that would fit your journal quite well. I know it's been a strange couple of years with Covid, but it'd be great to have some sort of process and reliability.

This review is taking longer than we'd usually expect but it's not completely out of the ordinary. Over the last ~1300 papers we've published our average review time is something around 3 months (http://www.theoj.org/joss-analytics/joss-submission-analytics.html#pre-reviewreview-time-over-time).
