### 2021-08-07 #3124 @nsaunier
I have left unanswered comments about the code for weeks now. I am not doing anything else till they are addressed. Thanks!---Nicolas Saunier, ing. DrSc / Ph.D.Professeur titulaire / Full ProfessorDépartement des génies civil, géologique et des mines (CGM) PolytechniqueMontréal CIRRELT, RRSR, CIRODD, IVADO & Chaire MOBILITÉhttp://nicolas.saunier.confins.netOn Aug. 7, 2021 02:48, Max Pumperla ***@***.***> wrote:
@jorgedch any chance we can find the time to address the above findings by @abhishekbajpayee above? Might be good to sync on a round of Q&A together.
@arfon @arokem is this the normal/expected process for a paper submission? I'm just asking because this smaller logodetect paper is a "test run" for more high-level projects that I'm running, and that would fit your journal quite well. I know it's been a strange couple of years with Covid, but it'd be great to have some sort of process and reliability.

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
