### 2021-08-07 #3124 @jorgedch
My apologies for the confusion to everyone, normally I take only a couple of days to address feedback. In the case of the feedback from @abhishekbajpayee, his comments are related to the installation process of the python package, and assumed that Max was going to address them as he implemented the packaging and is more familiar with it. Can you address this in the following days  @maxpumperla? Otherwise, I can make time on Thursday.

It would be very much appreciated if the feedback of the reviewers is done in parallel and not depending on the response to the feedback of other reviewers. Please let us know if you have any feedback on the algorithmic side of the submission @nsaunier, I can work on it in parallel to my co-author.
