+++ logodetect/classifiers/siamese.py
@@ -7,44 +7,43 @@ import torch
 from logodetect import classifiers
 from logodetect.augmenters import get_augmentations
 from logodetect.utils import clean_name, open_and_resize, image_to_gpu_tensor
-from constants import (
-    CLASSIFIER_ALG,
-    CLASSIFIER_DEVICE,
-    CLASSIFIER_WEIGHTS,
-    IMAGE_RESIZE,
-    MIN_CONFIDENCE,
-)
+from logodetect.constants import get_recognizer_config
 
 
 class Classifier:
     """Siamese Network classifier."""
 
-    def __init__(self, exemplar_paths, classifier_algo=None):
+    def __init__(self, exemplar_paths, classifier_algo=None, config: dict = None):
+        self.config = get_recognizer_config(config)
         # Define class variables:
         self.exemplars_imgs = None
         self.exemplars_brands = None
 
-        algo = classifier_algo if classifier_algo else CLASSIFIER_ALG
+        algo = classifier_algo if classifier_algo else self.config.get("CLASSIFIER_ALG")
 
         # Set the network to classify the detections:
         self._load_exemplars(exemplar_paths)
-        self.classifier = classifiers.get(algo)(CLASSIFIER_DEVICE, CLASSIFIER_WEIGHTS)
+        self.classifier = classifiers.get(algo)(
+            self.config.get("CLASSIFIER_DEVICE"), self.config.get("CLASSIFIER_WEIGHTS")
+        )
 
     def _load_exemplars(self, exemplars_paths):
         self.exemplars_imgs = []
         self.exemplars_brands = []
         for path in exemplars_paths:
             brand = clean_name(path)
-            image = open_and_resize(path, IMAGE_RESIZE)
+            image = open_and_resize(path, self.config.get("IMAGE_RESIZE"))
 
             # Store clean image:
-            image_gpu = image_to_gpu_tensor(image, CLASSIFIER_DEVICE)
+            image_gpu = image_to_gpu_tensor(image, self.config.get("CLASSIFIER_DEVICE"))
             self.exemplars_imgs.append(image_gpu)
             self.exemplars_brands.append(brand)
 
             # Store augmented image:
             for aug_image in get_augmentations(image):
-                aug_image_gpu = image_to_gpu_tensor(aug_image, CLASSIFIER_DEVICE)
+                aug_image_gpu = image_to_gpu_tensor(
+                    aug_image, self.config.get("CLASSIFIER_DEVICE")
+                )
                 self.exemplars_imgs.append((aug_image_gpu))
                 self.exemplars_brands.append(brand)
 
@@ -65,13 +64,13 @@ class Classifier:
         """
         comb_images = []
         for box in detections["boxes"]:
-            crop = image.crop(box).resize(IMAGE_RESIZE)
-            detection = image_to_gpu_tensor(crop, CLASSIFIER_DEVICE)
+            crop = image.crop(box).resize(self.config.get("IMAGE_RESIZE"))
+            detection = image_to_gpu_tensor(crop, self.config.get("CLASSIFIER_DEVICE"))
 
             for exemplar in self.exemplars_imgs:
                 comb_images.append(torch.cat((detection, exemplar), 1))
 
-        return torch.cat(comb_images).to(CLASSIFIER_DEVICE)
+        return torch.cat(comb_images).to(self.config.get("CLASSIFIER_DEVICE"))
 
     def _process_scores(self, comb_scores, detections):
         n_detections = len(detections["boxes"])
@@ -83,7 +82,7 @@ class Classifier:
             b = a + n_logos
             scores = comb_scores[a:b]
 
-            if (scores >= MIN_CONFIDENCE).any():
[…]
