### 2021-08-12 5b3a497
fix minor opencv issue, prep 1.0.2
