### 2021-08-12 dd055cd
fix minor opencv issue, prep 1.0.2
