+++ tests/test_utils.py
@@ -6,13 +6,13 @@ import os
 @pytest.mark.unit
 def test_open_and_resize():
     resize = (100, 100)
-    img = open_and_resize("./data/exemplars/3m.jpg", resize)
+    img = open_and_resize("./data/exemplars/pepsi_1.jpg", resize)
     assert img.size == resize
 
 
 @pytest.mark.unit
 def test_image_to_gpu():
-    img = open_and_resize("./data/exemplars/3m.jpg", (200, 200))
+    img = open_and_resize("./data/exemplars/pepsi_1.jpg", (200, 200))
     image_to_gpu_tensor(
         img, "cpu"
     )  # can't use cuda etc. here, as not every system will have it
@@ -20,15 +20,15 @@ def test_image_to_gpu():
 
 @pytest.mark.unit
 def test_clean_name():
-    file_name = "./data/exemplars/bankofamerica_text.jpg"
+    file_name = "./data/exemplars/redbull_1.jpg"
     name = clean_name(file_name)
-    assert name == "bankofamerica"
+    assert name == "redbull"
 
 
 @pytest.mark.unit
 def test_save_df():
     vectors = [[1, 2, 3], [4, 5, 6]]
-    files = ["./data/exemplars/bankofamerica_text.jpg", "./data/exemplars/3m.jpg"]
+    files = ["./data/exemplars/redbull_1.jpg", "./data/exemplars/pepsi_1.jpg"]
     save_df(vectors, files, "./test_df")
     assert os.path.exists("./test_df.pkl")
     os.remove("./test_df.pkl")
