+++ tests/test_detectors.py
@@ -6,6 +6,6 @@ import pytest
 @pytest.mark.unit
 def test_detectors():
     detector = Detector()
-    img = open_and_resize("./data/exemplars/3m.jpg", (200, 200))
+    img = open_and_resize("./data/exemplars/pepsi_1.jpg", (200, 200))
     detections = detector.predict(img)
     assert sorted(list(detections.keys())) == ["boxes", "brands", "labels", "scores"]
