+++ setup.py
@@ -11,22 +11,27 @@ setup(
     author="Jorge Davila Chacon",
     author_email="jorge@heldenkombinat.com",
     install_requires=[
-        "Cython>=0.29.15",
         "click>=7.1.1",
+        "Cython>=0.29.15",
+        "Flask>=1.1.1",
+        "Flask-Cors>=3.0.9",
+        "gunicorn>=20.0.4",
         "imgaug>=0.4.0",
-        "matplotlib>=3.2.0rc3",
+        "matplotlib>=2.2.5",
         "moviepy>=1.0.1",
-        "numpy>=1.18.1",
+        "numpy>=1.18.2",
         "opencv-python>=4.2.0.32",
-        "pandas>=1.0.1",
+        "pandas>=1.0.3",
+        "pytest>=6.2.4",
+        "scikit-image>=0.16.2",
+        "scikit-learn>=0.22.1",
         "scipy>=1.4.1",
-        "sklearn",
-        "torch>=1.4.0",
-        "torchvision>=0.5.0",
+        "torch>=1.9.0",
+        "torchvision>=0.10.0",
         "tqdm>=4.42.1",
     ],
     extras_require={
-        "tests": ["pytest", "pytest-pep8", "pytest-cov", "mock"],
+        "tests": ["pytest", "pytest-pep8", "pytest-cov", "mock", "pylint", "autopep8"],
         "dev": ["black", "pre-commit", "flask", "gunicorn"],
     },
     entry_points={"console_scripts": ["logodetect=logodetect.cli:cli"]},
