+++ requirements.txt
@@ -1,19 +1,21 @@
-Cython==0.29.15
-imgaug==0.4.0
+click>=7.1.1
+Cython>=0.29.15
+Flask>=1.1.1
+Flask-Cors>=3.0.9
+gunicorn>=20.0.4
+imgaug>=0.4.0
 matplotlib>=2.2.5
-moviepy==1.0.1
-numpy==1.18.2
-opencv-python==4.2.0.32
-pandas==1.0.3
-scipy==1.4.1
-scikit-learn==0.22.1
-torch==1.4.0
-torchvision==0.5.0
-tqdm==4.42.1
-gunicorn==20.0.4
-Flask==1.1.1
-flask-cors==3.0.9
-scikit-image==0.16.2
+moviepy>=1.0.1
+numpy>=1.18.2
+opencv-python>=4.2.0.32
+pandas>=1.0.3
+pytest>=6.2.4
+scikit-image>=0.16.2
+scikit-learn>=0.22.1
+scipy>=1.4.1
+torch>=1.9.0
+torchvision>=0.10.0
+tqdm>=4.42.1
 
 # To make VS code happy, not required for this package
 pylint
