### 2021-08-13 #3124 @jorgedch
Dear reviewers,

After some updates, currently fresh installs on Mac and Ubuntu are running the installation process and the tests as indicated in the README file. There are some things to improve on our side, but are not blockers. For example, the installation order in pip seems to produce an error as a pip package is required to build the wheel of another one in the requirements, and the installation order seems to generate a conflict some times. We are looking into it.

Also some errors are shown in the logs when running the server for the browser, but are not related to our package and make no problems for the user in the browser. Everything else is fixed now :-)

FYI:

- [Mac] Error shows first time running pip install -e ".[tests, dev]":
Failed building wheel for matplotlib

- [Mac] Second time running pip install -e ".[tests, dev]" it shows:
DEPRECATION: matplotlib was installed using the legacy 'setup.py install' method, because a wheel could not be built for it. A possible replacement is to fix the wheel build issue reported above. You can find discussion regarding this at https://github.com/pypa/pip/issues/8368.

- [Mac] Third time running pip install -e ".[tests, dev]" it works without problems. Possibly related to cython needed as a dependency to build matplotlib wheel, a classic problem with the installation order in pip, so it disappears after a few runs when it is installed. Maybe we can instruct people to install cython by hand before running setup.py?

- [Mac and Ubuntu] Error shows only first time running python app.py:
192.168.0.205 - - [13/Aug/2021 19:20:17] "GET /favicon.ico HTTP/1.1" 404 -
