### 2021-08-31 b24dd29
clarify installation, bump version
