### 2021-08-31 bb234f4
fix home folder for windows users, bump version
