### 2021-08-31 #3124 @maxpumperla
@nsaunier @abhishekbajpayee we've addressed your concerns and published `logodetect` version `1.1.2`. If you'd be so kind to have a look, that'd be much appreciated.

@nsaunier if there are further problems, it would be great to please link the GitHub issues you might be creating in the `logodetect` repo _to this thread_. We're getting a lot of GitHub notifications and it can be intractable to understand who is who across repos.

Also, thanks @arfon for the clarification. I just wanted to get a better picture to adjust expectations. I completely understand if things take their time.
