### 2021-11-21 #3124 @arokem
@szkafander : have you had a chance to review the paper and software?

@nsaunier : where do things stand on your end? Have all your concerns been addressed? I see that you have not checked all of the boxes in your check list yet.
