### 2021-11-22 #3124 @szkafander
@arokem I figured that you wanted to move on with this review since it has been pending for a while. This is my experience with the code so far:

Installation does not work on Windows. I could not find any indication in the docs about the supported operating systems so I assumed that it should work on Windows. I was able to install on Linux. Did not try on Mac.

The package works as intended on demo data. I noticed glitches on user data. I opened an issue in the package repo.

Otherwise the code is well-documented and complies with JOSS standards. The ML is pretty cool. I am sure that the glitches can be fixed easily.

I went ahead and checked all boxes in my review but I would suggest that the authors fix the issues before you accept the paper.
