### 2021-11-22 #3124 @jorgedch
Thank you for your feedback @szkafander.

[Installation on Windows]
Good catch, the package is intended to be used in Linux and MacOS, so we'll clarify this in the documentation and setup file.

[Detection accuracy]
Indeed, it is the case that the accuracy of the logo detections of the provided model is not very high. We describe in the README the following:

"additionally to sharing with you a couple of algorithms to **get started with one-shot logo-detection**, the aim of this project is to develop a flexible architecture to **facilitate the comparison of different algorithms** for one-shot object recognition." and "The models that we are including in the repo achieved a **reasonable performance after a few training epochs.** However, if you would like to **improve their performance** you can find pointers to various datasets in [6], which can be used in the training part of this project."

So the aim of the project is not to provide the user with a highly performant pre-trained model, but with a playground to allow a standardized comparison of one-shot algorithms. However, as we continue to refine the training of models for the detection of logos (and other classes of objects, such as traffic signs, faces, etc.), we are very happy to share those improved models later on with the community.

Thank you once again for your feedback @szkafander. I'll update the repo with the changes that you point out about the OS 👍
