### 2021-11-23 1e71c24
Update setup.py
Specify supported operating system
