### 2022-01-03 #3124 @arokem
Hello @nsaunier : have you had a chance to take another look at this? I see that several check-boxes are unchecked in your list. Could you provide some more comments, that the authors can address these issues? Thanks!

@jorgedch : have you had a chance to address the issues raised by @szkafander?
