### 2022-01-03 #3124 @jorgedch
Hi @arokem. Yes, I made the changes the same day of my previous reply. The setup file now specifies that the intended OS is Linux only: `Operating System :: POSIX :: Linux`. Concerning the comment about the accuracy, it is also clarified why the provided model is not performing accurately with some images. Last time we heard from one of the reviewers was almost half a year ago. What is the journal policy for waiting times?
