### 2022-01-25 #3124 @arokem
@conradsnicta @vchuravy @abuccts: would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [Logodetect: One-shot detection of logos in image and video data, a Python package for detecting logos in images and videos](https://github.com/Heldenkombinat/Logodetect).
