### 2022-04-06 #3124 @arfon
@maxpumperla @jorgedch – I'm going to do the second review here myself. If/as I find issues, I'll ask them as I go:

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

The paper doesn't really address this instead saying _'there is plenty of literature on the use of deep-learning for detecting logos'_. Would it be possible to add a couple of sentences to your paper saying something about other implementations/approaches compare to Logodetect?
