### 2022-04-06 #3124 @arfon
Installing from source on a fresh GitHub Codespaces (after running `pip install -r requirements.txt`), I get the following error when initially attempting to start the `logodetect` CLI tool (as per docs [here](https://github.com/Heldenkombinat/Logodetect#usage)):

```
ImportError: libGL.so.1: cannot open shared object file: No such file or directory
```

I fixed this with `apt-get install libgl1`. Should `libgl1` included as part of the [`opencv` requirements](https://github.com/Heldenkombinat/Logodetect/blob/master/requirements.txt#L10)? If so, it doesn't seem to be 🤷 ?
