### 2022-04-06 #3124 @arfon
Running the test examples, the image-based detector seems to do a worse job than the video detector. Is this to be expected?

![test_image_small_output](https://user-images.githubusercontent.com/4483/161925429-bb4376b3-4e53-4978-9d29-c2de252b7e18.png)

https://user-images.githubusercontent.com/4483/161925438-3fc36e05-6e93-4697-b60a-0f20ec6e11a9.mp4
