### 2022-04-06 #3124 @arfon
OK, so in summary, I think the following issues need to be addressed for me to sign off on this review:

- Installation instructions: Resolving the potential issue around `libGL` dependency [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089949375)].
- Addressing feedback on 'State of the field' in the paper [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089925666)]
- Missing community guidelines [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089969176)]
