### 2022-04-24 c295ac2
Clarified "how this software compares to other commonly-used packages"
