+++ SUPPORT.md
@@ -1,0 +1,29 @@
+# Support
+
+Thank you for your interest in Logodetect!
+If you have any questions, please contact us at [logodetect@heldenkombinat.com][logodetect@heldenkombinat.com]
+Either by reporting a bug, by suggesting a feature or by submitting a pull request,
+you can help us improve the project.
+
+The `README` in the root of the repository should contains the
+project documentation. If you cannot find the documentation you're
+looking for, please file a GitHub issue with details of what
+you'd like to see documented.
+
+
+## Code of Conduct
+
+By participating in this project, you are expected to uphold
+our [Code of Conduct][support/CODE_OF_CONDUCT.md].
+
+
+## Issues
+
+Please make use of the following templates for your reports and requests:
+- [`BUG_REPORT.md`][support/BUG_REPORT.md]
+- [`FEATURE_REQUEST.md`][support/FEATURE_REQUEST.md]
+
+
+## Contributions
+
+For contributions, please follow the indications in [`PULL_REQUEST.md`][support/PULL_REQUEST.md].
