+++ support/FEATURE_REQUEST.md
@@ -1,0 +1,22 @@
+---
+name: Feature request
+about: Suggest a new feature or improvements to an existing one
+labels: "✨ goal: improvement, 🚦 status: awaiting triage, 💻 aspect: code, 🟩 priority: low"
+title: "[Feature] <Replace this with actual title>"
+---
+
+## Problem
+<!-- Describe a problem solved by this feature; or delete the section entirely. -->
+
+## Description
+<!-- Describe the feature and how it solves the problem. -->
+
+## Alternatives
+<!-- Describe any alternative solutions or features you have considered. How is this feature better? -->
+
+## Additional context
+<!-- Add any other context about the feature here; or delete the section entirely. -->
+
+## Implementation
+<!-- Replace the [ ] with [x] to check the box. -->
+- [ ] I would be interested in implementing this feature.
