### 2022-04-24 #3124 @jorgedch
Hi @arfon,

Thank you once again for your feedback, we have addressed it as follows:

> OK, so in summary, I think the following issues need to be addressed for me to sign off on this review:
>
> * Installation instructions: Resolving the potential issue around `libGL` dependency [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089949375)].

Done. We added the clarification that the library has a dependency on libgl1 and the command to install it.

> * Addressing feedback on 'State of the field' in the paper [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089925666)]

We have clarified the distinction of our library as follows:

"we also provide a modified ResNet [2] for object-recognition that directly takes the exemplars and the detections from the first stage and predicts if both belong to the same sub-class. Similarly to [3], this network infers a distance metric after being trained with examples of different sub-classes, **but instead of sharing the same weights and processing each input in a separate pass as in [4], it concatenates both inputs and processes them in one pass**. This concept follows more closely the architecture proposed by [5], where the assumption is that the exemplars often have more high-frequency components than the detections, and therefore the model can increase its accuracy by learning a separate set of weights for each input. **However, our proposed architecture splits the detection and classification tasks in two separate stages, which allows the use, and comparison, of different classifiers for the second stage.**"

> * Clarifying if the test behaviour with images and videos is expected [link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089962647)

Yes, this behavior is expected. The baseline model uses a distance threshold to decide whether a detected logo belongs to one of the brands that the user is searching for. For this reason, in different video frames the classifier may behave differently between images that are fairly similar. However, the baseline model is the simplest approach and the reference model, as well as other models that the user wants to test, improve over the baseline.

> * Missing community guidelines [[link](https://github.com/openjournals/joss-reviews/issues/3124#issuecomment-1089969176)]

We have added now documentation for support, issues and contributions.

I believe this addresses the feedback points. Looking forward to the next steps :) All the best!
