### 2022-04-24 #3124 @arfon
OK thanks @jorgedch! I've found one minor issue here that could do with fixing: https://github.com/Heldenkombinat/Logodetect/pull/40

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
