### 2022-04-25 #3124 @arfon
@maxpumperla – your paper is now accepted and published in JOSS :zap::rocket::boom:. Sorry it took a while longer than expected 😅
